requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.2
pyparsing==2.4.7
pyspark==3.1.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
werkzeug==2.2.2
